implement StateTest
